refactor(town): extract shop helpers to worldgen/town/shops_core.js and wire into town_gen.js (loadShopDefs, scheduleFromData, minutesOfDay, shuffle, chanceFor)
