+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/worldgen/town_gen.js
@@ -24,10 +24,11 @@
 
 import { getGameData, getMod, getRNGUtils } from "../utils/access.js";
 import { getTownBuildingConfig, getInnSizeConfig, getCastleKeepSizeConfig, getTownPopulationTargets } from "./town/config.js";
-import { buildBaseTown, buildPlaza, carveBuildingRect, placeCastleKeep, buildInnAndMarkTavern, isAreaClearForBuilding, findBuildingsOverlappingRect, layoutCandidateDoors, layoutEnsureDoor, layoutGetExistingDoor } from "./town/layout_core.js";
+import { buildBaseTown, buildPlaza, carveBuildingRect, placeCastleKeep, buildInnAndMarkTavern, isAreaClearForBuilding, findBuildingsOverlappingRect, layoutCandidateDoors, layoutEnsureDoor, layoutGetExistingDoor, placePlazaPrefabStrict } from "./town/layout_core.js";
 import { buildOutdoorMask, repairBuildingPerimeters, placeWindowsOnAll } from "./town/windows.js";
 import { addProp, addSignNear, addShopSignInside, dedupeShopSigns, dedupeWelcomeSign, cleanupDanglingProps } from "./town/signs.js";
 import { spawnGateGreeters, enforceGateNPCLimit, populateTownNpcs } from "./town/npcs_bootstrap.js";
+import { minutesOfDay, scheduleFromData, loadShopDefs, shopLimitBySize, chanceFor, shuffleInPlace } from "./town/shops_core.js";
 
 function inBounds(ctx, x, y) {
   try {
@@ -844,43 +845,8 @@ function generate(ctx) {
     } catch (_) {}
   })();
 
-  // Data-first shop selection: use GameData.shops when available
-  
-  function minutesOfDay(ctx, h, m = 0) {
-    try {
-      if (ctx && ctx.ShopService && typeof ctx.ShopService.minutesOfDay === "function") {
-        return ctx.ShopService.minutesOfDay(h, m, 24 * 60);
-      }
-    } catch (_) {}
-    return ((h | 0) * 60 + (m | 0)) % (24 * 60);
-  }
-  function scheduleFromData(row) {
-    if (!row) return { openMin: minutesOfDay(ctx, 8), closeMin: minutesOfDay(ctx, 18), alwaysOpen: false };
-    if (row.alwaysOpen) return { openMin: 0, closeMin: 0, alwaysOpen: true };
-    const o = parseHHMM(row.open);
-    const c = parseHHMM(row.close);
-    if (o == null || c == null) return { openMin: minutesOfDay(ctx, 8), closeMin: minutesOfDay(ctx, 18), alwaysOpen: false };
-    return { openMin: o, closeMin: c, alwaysOpen: false };
-  }
-
-  // Shop definitions: disable data-assigned shops only when strict prefabs are available
-  const GD9 = getGameData(ctx);
-  let shopDefs = strictNow
-    ? []
-    : ((GD9 && Array.isArray(GD9.shops)) ? GD9.shops.slice(0) : [
-        { type: "inn", name: "Inn", alwaysOpen: true },
-        { type: "blacksmith", name: "Blacksmith", open: "08:00", close: "17:00" },
-        { type: "apothecary", name: "Apothecary", open: "09:00", close: "18:00" },
-        { type: "armorer", name: "Armorer", open: "08:00", close: "17:00" },
-        { type: "trader", name: "Trader", open: "08:00", close: "18:00" },
-      ]);
-  try {
-    const idxInn = shopDefs.findIndex(d => String(d.type || "").toLowerCase() === "inn" || /inn/i.test(String(d.name || "")));
-    if (idxInn > 0) {
-      const innDef = shopDefs.splice(idxInn, 1)[0];
-      shopDefs.unshift(innDef);
-    }
-  } catch (_) {}
+  // Data-first shop selection: use GameData.shops when available (helpers in town/shops_core.js)
+  let shopDefs = loadShopDefs(ctx, strictNow);
 
   // Score buildings by distance to plaza and assign shops to closest buildings
   const scored = buildings.map(b => ({ b, d: Math.abs((b.x + ((b.w / 2))) - plaza.x) + Math.abs((b.y + ((b.h / 2))) - plaza.y) }));
@@ -888,34 +854,10 @@ function generate(ctx) {
   // Track largest building by area for assigning the inn
   
 
-  // Vary number of shops by town size
-  function shopLimitBySize(sizeKey) {
-    if (sizeKey === "small") return 3;
-    if (sizeKey === "city") return 8;
-    return 5; // big
-  }
+  // Vary number of shops by town size (helper in town/shops_core.js)
   const limit = Math.min(scored.length, shopLimitBySize(townSize));
 
-  // Deterministic sampling helpers for shop presence
-  function chanceFor(def, sizeKey) {
-    try {
-      const c = def && def.chanceBySize ? def.chanceBySize : null;
-      if (c && typeof c[sizeKey] === "number") {
-        const v = c[sizeKey];
-        return (v < 0 ? 0 : (v > 1 ? 1 : v));
-      }
-    } catch (_) {}
-    // Defaults if not specified in data
-    if (sizeKey === "city") return 0.75;
-    if (sizeKey === "big") return 0.60;
-    return 0.50; // small
-  }
-  function shuffleInPlace(arr) {
-    for (let i = arr.length - 1; i > 0; i--) {
-      const j = Math.floor(rng() * (i + 1));
-      const t = arr[i]; arr[i] = arr[j]; arr[j] = t;
-    }
-  }
+  // Deterministic sampling helpers for shop presence (helpers in town/shops_core.js)
 
   // Build shop selection: Inn always included, others sampled by chanceBySize (dedup by type)
   let innDef = null;
@@ -930,10 +872,10 @@ function generate(ctx) {
   let sampled = [];
   for (const d of candidateDefs) {
     const ch = chanceFor(d, townSize);
-    if (rng() < ch) sampled.push(d);
+    if (rng() &lt; ch) sampled.push(d);
   }
   // Shuffle and cap, but avoid duplicate types within a single town
-  shuffleInPlace(sampled);
+  shuffleInPlace(sampled, rng);
   const restCap = Math.max(0, limit - (innDef ? 1 : 0));
   const finalDefs = [];
   const usedTypes = new Set();
@@ -1022,7 +964,7 @@ function generate(ctx) {
     } else {
       door = ensureDoor(b);
     }
-    const sched = scheduleFromData(def);
+    const sched = scheduleFromData(ctx, def);
     const name = def.name || def.type || "Shop";
 
     // inside near door
